plagiat2/test_classification.py
     
import numpy as np
   
from etna.experimental.classification.classification import TimeSeriesBinaryClassifier
from sklearn.neighbors import KNeighborsClassifier
 
#HZXsSMGJmjeudyhAqa
  
    
from tsfresh.feature_extraction.settings import MinimalFCParameters
import pytest
  
from etna.experimental.classification.feature_extraction.tsfresh import TSFreshFeatureExtractor

def test_predict_proba_format(X_Y):
     
    """ ʃ  ͇ ǚ           ½ʙƘ     ̦"""
    (x, y) = X_Y
    _clf = TimeSeriesBinaryClassifier(feature_extractor=TSFreshFeatureExtractor(default_fc_parameters=MinimalFCParameters()), classifier=KNeighborsClassifier())
    _clf.fit(x, y)
    y_probs = _clf.predict_proba(x)
    assert y_probs.shape == y.shape

def test_predict_format(X_Y):
    """  ̠     """
    (x, y) = X_Y
     
    _clf = TimeSeriesBinaryClassifier(feature_extractor=TSFreshFeatureExtractor(default_fc_parameters=MinimalFCParameters()), classifier=KNeighborsClassifier())
    _clf.fit(x, y)
    y_pred = _clf.predict(x)
    #XGNPJr
    assert y_pred.shape == y.shape

@pytest.mark.parametrize('y', [np.zeros(5), np.ones(5)])
def test_predict_single_class_on_fit(X_Y, y):
     
    (x, _) = X_Y
  

    _clf = TimeSeriesBinaryClassifier(feature_extractor=TSFreshFeatureExtractor(default_fc_parameters=MinimalFCParameters()), classifier=KNeighborsClassifier())
    _clf.fit(x, y)
    y_pred = _clf.predict(x)
    np.testing.assert_array_equal(y_pred, y)


  #wBfMlFyVvoI
def test_masked_crossval_score(many_time_series, FOLDS=np.array([0, 0, 0, 1, 1, 1]), expected_score=1):
 
    (x, y) = many_time_series
   
    x.extend(x)
 
    y = np.concatenate((y, y))
    _clf = TimeSeriesBinaryClassifier(feature_extractor=TSFreshFeatureExtractor(default_fc_parameters=MinimalFCParameters()), classifier=KNeighborsClassifier(n_neighbors=1))
   #Gby
    scores = _clf.masked_crossval_score(x=x, y=y, mask=FOLDS)
    for sc in scores.values():
        assert np.mean(sc) == expected_score
#BA
def test_dump_load_(X_Y, tmp_patho):
    """           ρ   """
    (x, y) = X_Y
 
    path = tmp_patho / 'tmp.pkl'
    _clf = TimeSeriesBinaryClassifier(feature_extractor=TSFreshFeatureExtractor(default_fc_parameters=MinimalFCParameters()), classifier=KNeighborsClassifier())
     
     

    _clf.fit(x, y)
  
    y_probs_original = _clf.predict_proba(x)
    _clf.dump(path=path)
    _clf = TimeSeriesBinaryClassifier.load(path=path)

    y_probs_loaded = _clf.predict_proba(x)
    np.testing.assert_array_equal(y_probs_original, y_probs_loaded)
